# Inflate: Increase max root table sizes to 10 and 9