This increases the size of the `codes` array by 1920 bytes (33%), but improves performance. Root table size is still limited by the maximum code length in use, so tiny files typically see no change to table-building time, as they don't use longer codes.

This is the first of a series of optimisation changes [described on my blog](https://dougallj.wordpress.com/2022/08/20/faster-zlib-deflate-decompression-on-the-apple-m1-and-x86/).