@nmoinvaz 2022-08-23
# Silesia-small.tar 15MB

## HEAD https://github.com/zlib-ng/zlib-ng/commit/89763032d57e3da5301f4cd6e1f363e7a1f85f02
```
OS: Windows 10 10.0.22000 AMD64
CPU: Intel64 Family 6 Model 158 Stepping 10, GenuineIntel
Tool: ../zlib-ng/build-inf-develop/Release/minigzip.exe
Levels: 0-9
Runs: 70         Trim worst: 40

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.008%      0.015/0.016/0.016/0.000        0.014/0.017/0.025/0.004       15,737,543
 1     54.166%      0.086/0.088/0.090/0.001        0.043/0.053/0.058/0.006        8,523,732
 2     43.871%      0.141/0.145/0.148/0.002        0.044/0.056/0.059/0.004        6,903,609
 3     42.387%      0.195/0.202/0.212/0.005        0.043/0.055/0.058/0.005        6,670,099
 4     41.647%      0.218/0.220/0.224/0.002        0.042/0.048/0.056/0.006        6,553,723
 5     41.216%      0.226/0.230/0.234/0.002        0.042/0.052/0.056/0.006        6,485,938
 6     41.037%      0.262/0.266/0.271/0.002        0.041/0.042/0.043/0.001        6,457,776
 7     40.778%      0.329/0.335/0.344/0.004        0.042/0.052/0.057/0.006        6,416,919
 8     40.704%      0.421/0.426/0.435/0.004        0.042/0.046/0.056/0.005        6,405,244
 9     40.409%      0.494/0.500/0.505/0.003        0.042/0.047/0.056/0.006        6,358,951

 avg1  48.622%                        0.243                          0.047
 avg2  54.025%                        0.270                          0.052
 tot                                 72.783                         14.024       76,513,534
```
## PR https://github.com/zlib-ng/zlib-ng/pull/1332/commits/60216d0837702e068a7484363bb85bfea9bd465f
```
OS: Windows 10 10.0.22000 AMD64
CPU: Intel64 Family 6 Model 158 Stepping 10, GenuineIntel
Tool: ../zlib-ng/build-inf-tables/Release/minigzip.exe
Levels: 0-9
Runs: 70         Trim worst: 40

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.008%      0.015/0.023/0.026/0.005        0.014/0.020/0.025/0.005       15,737,543
 1     54.166%      0.086/0.087/0.088/0.001        0.043/0.054/0.058/0.006        8,523,732
 2     43.871%      0.141/0.145/0.147/0.001        0.043/0.054/0.057/0.004        6,903,609
 3     42.387%      0.196/0.201/0.211/0.003        0.041/0.049/0.056/0.007        6,670,099
 4     41.647%      0.219/0.222/0.231/0.003        0.040/0.044/0.053/0.004        6,553,723
 5     41.216%      0.227/0.236/0.242/0.006        0.040/0.047/0.054/0.006        6,485,938
 6     41.037%      0.263/0.274/0.279/0.005        0.039/0.041/0.052/0.003        6,457,776
 7     40.778%      0.330/0.343/0.348/0.006        0.041/0.052/0.055/0.004        6,416,919
 8     40.704%      0.417/0.433/0.436/0.004        0.039/0.043/0.053/0.005        6,405,244
 9     40.409%      0.498/0.508/0.512/0.004        0.040/0.046/0.054/0.006        6,358,951

 avg1  48.622%                        0.247                          0.045
 avg2  54.025%                        0.274                          0.050
 tot                                 74.110                         13.503       76,513,534
```

# Silesia.tar 200MB

## HEAD https://github.com/zlib-ng/zlib-ng/commit/89763032d57e3da5301f4cd6e1f363e7a1f85f02
```
OS: Windows 10 10.0.22000 AMD64
CPU: Intel64 Family 6 Model 158 Stepping 10, GenuineIntel
Tool: ../zlib-ng/build-inf-develop/Release/minigzip.exe
Levels: 0-9
Runs: 70         Trim worst: 40

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.008%      0.146/0.148/0.150/0.001        0.123/0.125/0.126/0.001      211,973,953
 1     44.409%      0.916/0.922/0.928/0.003        0.464/0.469/0.473/0.003       94,127,290
 2     35.518%      1.550/1.565/1.570/0.005        0.484/0.489/0.493/0.003       75,282,961
 3     33.882%      2.173/2.186/2.191/0.005        0.461/0.467/0.472/0.003       71,816,478
 4     33.174%      2.429/2.438/2.443/0.003        0.450/0.457/0.465/0.004       70,315,668
 5     32.660%      2.531/2.540/2.547/0.005        0.452/0.457/0.461/0.003       69,225,542
 6     32.508%      2.941/2.957/2.966/0.006        0.448/0.454/0.460/0.004       68,902,222
 7     32.255%      3.941/3.954/3.963/0.007        0.447/0.457/0.460/0.003       68,366,800
 8     32.167%      6.219/6.238/6.251/0.009        0.451/0.457/0.461/0.003       68,180,776
 9     31.887%      6.670/6.696/6.707/0.009        0.447/0.455/0.458/0.003       67,586,442

 avg1  40.847%                        2.964                          0.429
 avg2  45.385%                        3.294                          0.476
 tot                                889.317                        128.622      865,778,132
```
## PR https://github.com/zlib-ng/zlib-ng/pull/1332/commits/60216d0837702e068a7484363bb85bfea9bd465f
```
OS: Windows 10 10.0.22000 AMD64
CPU: Intel64 Family 6 Model 158 Stepping 10, GenuineIntel
Tool: ../zlib-ng/build-inf-tables/Release/minigzip.exe
Levels: 0-9
Runs: 70         Trim worst: 40

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.008%      0.145/0.148/0.149/0.001        0.122/0.125/0.126/0.001      211,973,953
 1     44.409%      0.917/0.922/0.927/0.003        0.465/0.470/0.474/0.002       94,127,290
 2     35.518%      1.561/1.567/1.573/0.003        0.456/0.463/0.469/0.003       75,282,961
 3     33.882%      2.175/2.187/2.195/0.006        0.436/0.442/0.449/0.004       71,816,478
 4     33.174%      2.429/2.441/2.449/0.005        0.425/0.434/0.440/0.004       70,315,668
 5     32.660%      2.527/2.541/2.549/0.005        0.426/0.430/0.435/0.003       69,225,542
 6     32.508%      2.944/2.959/2.969/0.006        0.420/0.427/0.432/0.004       68,902,222
 7     32.255%      3.936/3.957/3.965/0.007        0.426/0.431/0.435/0.003       68,366,800
 8     32.167%      6.221/6.237/6.249/0.008        0.426/0.431/0.434/0.002       68,180,776
 9     31.887%      6.665/6.694/6.709/0.011        0.423/0.433/0.435/0.003       67,586,442

 avg1  40.847%                        2.965                          0.409
 avg2  45.385%                        3.295                          0.454
 tot                                889.643                        122.584      865,778,132
```