@Dead2 2022-09-23
Sorry for taking so long to accept this PR, I have a huge amount of things to get done before winter, and I have been trying to get [Stabilizer](https://github.com/Dead2/stabilizer) fixed up to a point where it is possible to use it for benchmarking smaller performance changes like some of yours. It is still only partially working, but it does make the benchmarks a little more accurate even with only the SZ_LOWER and SZ_HEAP features enabled.

## Baseline with Stabilizer on X86_64
```
 Tool: minigzip   Levels: 1-9
 Runs: 45         Trim worst: 5

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     44.409%      1.053/1.085/1.107/0.012        0.562/0.574/0.585/0.006       94,127,290
 2     35.518%      2.072/2.099/2.126/0.017        0.587/0.601/0.611/0.006       75,282,961
 3     33.882%      2.597/2.652/2.682/0.020        0.559/0.574/0.581/0.006       71,816,478
 4     33.174%      2.927/3.023/3.091/0.037        0.537/0.557/0.570/0.008       70,315,668
 5     32.660%      3.171/3.250/3.303/0.032        0.525/0.550/0.561/0.008       69,225,542
 6     32.508%      3.707/3.826/3.930/0.061        0.529/0.545/0.554/0.007       68,902,222
 7     32.255%      5.108/5.274/5.483/0.093        0.531/0.543/0.551/0.005       68,366,800
 8     32.167%      7.799/7.994/8.323/0.145        0.523/0.543/0.552/0.006       68,180,776
 9     31.887%      8.229/8.359/8.531/0.075        0.529/0.545/0.558/0.006       67,586,442

 avg1  34.273%                        4.173                          0.559
 tot                               1502.441                        201.215      653,804,179

 Tool: minigzip   Levels: 1-9
 Runs: 45         Trim worst: 5

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     44.409%      1.064/1.088/1.116/0.015        0.556/0.571/0.580/0.006       94,127,290
 2     35.518%      2.034/2.088/2.126/0.024        0.587/0.602/0.613/0.007       75,282,961
 3     33.882%      2.597/2.650/2.689/0.023        0.566/0.575/0.583/0.004       71,816,478
 4     33.174%      2.908/3.004/3.059/0.035        0.544/0.558/0.568/0.007       70,315,668
 5     32.660%      3.134/3.239/3.317/0.045        0.536/0.549/0.561/0.007       69,225,542
 6     32.508%      3.708/3.799/3.905/0.053        0.531/0.546/0.556/0.007       68,902,222
 7     32.255%      5.095/5.264/5.436/0.075        0.530/0.546/0.555/0.007       68,366,800
 8     32.167%      7.699/7.931/8.139/0.105        0.530/0.543/0.552/0.005       68,180,776
 9     31.887%      8.214/8.384/8.598/0.110        0.524/0.544/0.557/0.007       67,586,442

 avg1  34.273%                        4.161                          0.559
 tot                               1497.875                        201.359      653,804,179

   text    data     bss     dec     hex filename
 153746    1296      48  155090   25dd2 libz-ng.so.2
```

## PR 1332 with Stabilizer on x86_64
```
 Tool: minigzip   Levels: 1-9
 Runs: 45         Trim worst: 5

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     44.409%      1.053/1.089/1.120/0.014        0.559/0.572/0.582/0.007       94,127,290
 2     35.518%      2.054/2.088/2.117/0.017        0.560/0.572/0.582/0.006       75,282,961
 3     33.882%      2.606/2.645/2.702/0.023        0.528/0.542/0.554/0.006       71,816,478
 4     33.174%      2.946/3.023/3.101/0.035        0.514/0.526/0.536/0.006       70,315,668
 5     32.660%      3.171/3.256/3.373/0.050        0.501/0.517/0.528/0.006       69,225,542
 6     32.508%      3.681/3.799/3.907/0.058        0.498/0.512/0.523/0.006       68,902,222
 7     32.255%      5.134/5.245/5.356/0.061        0.496/0.512/0.523/0.006       68,366,800
 8     32.167%      7.774/7.952/8.067/0.077        0.484/0.509/0.520/0.007       68,180,776
 9     31.887%      8.227/8.373/8.587/0.110        0.502/0.517/0.532/0.008       67,586,442

 avg1  34.273%                        4.163                          0.531
 tot                               1498.834                        191.149      653,804,179

 Tool: minigzip   Levels: 1-9
 Runs: 45         Trim worst: 5

 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     44.409%      1.065/1.089/1.118/0.014        0.556/0.572/0.586/0.007       94,127,290
 2     35.518%      2.064/2.093/2.128/0.017        0.547/0.571/0.581/0.008       75,282,961
 3     33.882%      2.587/2.658/2.713/0.031        0.525/0.543/0.554/0.008       71,816,478
 4     33.174%      2.936/3.024/3.095/0.040        0.512/0.525/0.536/0.006       70,315,668
 5     32.660%      3.160/3.254/3.320/0.039        0.499/0.515/0.529/0.007       69,225,542
 6     32.508%      3.649/3.804/3.910/0.064        0.495/0.513/0.527/0.007       68,902,222
 7     32.255%      5.138/5.248/5.343/0.058        0.494/0.512/0.523/0.007       68,366,800
 8     32.167%      7.774/7.928/8.157/0.089        0.500/0.512/0.523/0.006       68,180,776
 9     31.887%      8.207/8.364/8.656/0.098        0.502/0.515/0.528/0.006       67,586,442

 avg1  34.273%                        4.163                          0.531
 tot                               1498.512                        191.095      653,804,179

   text    data     bss     dec     hex filename
 153746    1296      48  155090   25dd2 libz-ng.so.2
 ```
 
This looks like a very solid 5% speedup on inflate in my x86_64 tests.
As expected there are no changes in code or data sizes.

Btw, Stabilizer when fully functional should minimize the unrelated performance variance we see for averaged deflate numbers here, but due to some bugs causing crashes in zlib-ng the most important features can unfortunately not be enabled yet. If anyone has experience with LLVM-pass/LLVM-ir and is able to help with fixing this bug, a lot of people it would be appreciated by a lot of people.